companion: Move CompanionAPI to separate module
Relates to #1302
